DB Publisher: if possible, dont lock when sending
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@10932 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ BonjourPublisher.m
@@ -395,7 +395,7 @@ extern const char *GetPrivateIP();
                             {
                                 NSFileHandle *dbFileHandle = [NSFileHandle fileHandleForReadingAtPath: databasePath];
                                 NSLog( @"split DB file reading");
-                                int length = 0;
+                                NSUInteger length = 0;
                                 do
                                 {
                                     NSAutoreleasePool *arp = [[NSAutoreleasePool alloc] init];
